Fix device display off-by-one error
There has long been an off-by-one error in the params, where nbin does not actually target the device displayed in the script menu.  Instead it targets the device at index n - 1 when displaying the device at index n.

This bug has caused a lot of confused questions in the forum.